Renamed some commands and added audio commands
Part of the goal of this application is to make it easier to write scripts for
visual novels made with Ren'Py. Part of that is the contraction of common use
functions. The floowing functions are rarely used in a script and it didn't make
sence to contract them, so they have been renamed to make them more descriptive:

- `:l` -> `:line`
- `:p` -> `:character`
- `:m` -> `:choice`

The following commands for controling audio and music have been added:

- `:p`
- `:pm`
- `:ps`
- `:pv`
- `:pa`
- `:v`
- `:q`
- `:stop`

Also updated test file to accomodate these changes. This closes #8